+++ permission/src/commonMain/kotlin/tech/kotlinlang/permission/result/NotificationPermissionResult.kt
@@ -1,0 +1,7 @@
+package tech.kotlinlang.permission.result
+
+sealed interface NotificationPermissionResult {
+    data object Granted: NotificationPermissionResult
+    data object NotAllowed: NotificationPermissionResult
+    data object Denied: NotificationPermissionResult
+}